+++ heylogs-cli/src/main/java/nbbrd/heylogs/cli/ScanCommand.java
@@ -15,7 +15,7 @@ import java.util.concurrent.Callable;
 import static internal.heylogs.cli.MarkdownInputSupport.newMarkdownInputSupport;
 import static nbbrd.console.picocli.text.TextOutputSupport.newTextOutputSupport;
 
-@Command(name = "scan")
+@Command(name = "scan", description = "Summarize changelog content.")
 public final class ScanCommand implements Callable<Void> {
 
     @CommandLine.Mixin