+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-10-20
+
 ### Added
 
 - Add optional semantic versioning rule [#22](https://github.com/nbbrd/heylogs/issues/22)
@@ -54,7 +56,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/heylogs/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/nbbrd/heylogs/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/nbbrd/heylogs/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/nbbrd/heylogs/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/nbbrd/heylogs/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/nbbrd/heylogs/compare/v0.2.0...v0.3.0